batch: Move baseline reference positions
The batch merge path has baseline-coordinate fallback edits for cases where
structural anchors do not map selected changes onto the target file.

The project still keeps the coordinate proof for recorded baseline
references inside the fallback edit module. That makes the fallback module
own reference verification plus edit assembly, even though those concerns
change for different reasons.

This commit addresses that split by moving insertion and absence position
lookup into a baseline_reference_positions module. Baseline fallback edits
now call that public lookup API through role-named collaborators.

A following tests commit will pin the ownership boundary so future fallback
work keeps the reference-position API in its module.
