+++ src/git_stage_batch/batch/baseline_edits.py
@@ -11,6 +11,10 @@ from ..core.line_selection import LineRanges, LineSelection, coerce_line_ranges
 from ..exceptions import MergeError as _MergeError
 from ..i18n import _
 from ..utils.text import normalize_line_endings
+from .baseline_reference_positions import (
+    baseline_reference_absence_position as _find_baseline_absence_position,
+    baseline_reference_insertion_position as _find_baseline_insertion_position,
+)
 from .line_mapping import LineMapping
 from .merge_candidates import MergeResolution as _MergeResolution
 
@@ -59,113 +63,6 @@ def _sequence_digest(lines: Sequence[bytes]) -> str:
     return hashlib.sha256(b"".join(lines)).hexdigest()[:12]
 
 
-def _line_payload_for_reference_match(content: Any) -> bytes:
-    """Normalize one line for insertion-boundary identity checks."""
-    normalized = normalize_line_endings(bytes(content))
-    if normalized.endswith(b"\n"):
-        return normalized[:-1]
-    return normalized
-
-
-def _reference_line_matches(
-    target_line: bytes,
-    reference_content: bytes | None,
-) -> bool:
-    if reference_content is None:
-        return False
-    return (
-        _line_payload_for_reference_match(target_line)
-        == _line_payload_for_reference_match(reference_content)
-    )
-
-
-def _baseline_reference_insertion_position(
-    reference: Any,
-    working_lines: Sequence[bytes],
-) -> int | None:
-    """Return the proven insertion position for a baseline reference."""
-    if reference is None or not getattr(reference, "has_after_line", False):
-        return None
-
-    after_line = getattr(reference, "after_line", None)
-    position = after_line or 0
-    if position < 0 or position > len(working_lines):
-        return None
-
-    verified_boundary = False
-    if after_line is not None:
-        if after_line < 1 or after_line > len(working_lines):
-            return None
-        if not _reference_line_matches(
-            working_lines[after_line - 1],
-            getattr(reference, "after_content", None),
-        ):
-            return None
-        verified_boundary = True
-
-    if getattr(reference, "has_before_line", False):
-        before_line = getattr(reference, "before_line", None)
-        if before_line is None:
-            if position != len(working_lines):
-                return None
-            verified_boundary = True
-        else:
-            if position >= len(working_lines):
-                return None
-            if not _reference_line_matches(
-                working_lines[position],
-                getattr(reference, "before_content", None),
-            ):
-                return None
-            verified_boundary = True
-
-    if not verified_boundary:
-        return None
-    return position
-
-
-def _baseline_reference_absence_position(
-    reference: Any,
-    working_lines: Sequence[bytes],
-    sequence_length: int,
-) -> int | None:
-    """Return the proven removal position for a baseline reference."""
-    if reference is None or not getattr(reference, "has_after_line", False):
-        return None
-
-    after_line = getattr(reference, "after_line", None)
-    position = after_line or 0
-    if position < 0 or position + sequence_length > len(working_lines):
-        return None
-
-    after_content = getattr(reference, "after_content", None)
-    if after_line is not None and after_content is not None:
-        if after_line < 1 or after_line > len(working_lines):
-            return None
-        if not _reference_line_matches(
-            working_lines[after_line - 1],
-            after_content,
-        ):
-            return None
-
-    if getattr(reference, "has_before_line", False):
-        before_line = getattr(reference, "before_line", None)
-        before_position = position + sequence_length
-        if before_line is None:
-            if before_position != len(working_lines):
-                return None
-        else:
-            if before_position >= len(working_lines):
-                return None
-            if not _reference_line_matches(
-                working_lines[before_position],
-                getattr(reference, "before_content", None),
-            ):
-                return None
-
-    return position
-
-
 def _line_sequences_equal(
     left: Sequence[bytes],
     right: Sequence[bytes],
@@ -202,7 +99,7 @@ def _baseline_removal_edit(
         normalize_line_endings(line)
         for line in claim.content_lines
     ]
-    position = _baseline_reference_absence_position(
+    position = _find_baseline_absence_position(
         claim.baseline_reference,
         working_lines,
         len(forbidden_sequence),
@@ -225,7 +122,7 @@ def _replacement_origin_absence_bounds(
     if type(old_line_count) is not int or old_line_count <= 0:
         return None
 
-    position = _baseline_reference_absence_position(
+    position = _find_baseline_absence_position(
         origin.baseline_reference,
         working_lines,
         old_line_count,
@@ -590,7 +487,7 @@ def try_apply_baseline_replacement_units(
             if claimed_line < 1 or claimed_line > len(source_lines):
                 return None
             reference = claimed_line_references.get(claimed_line)
-            position = _baseline_reference_insertion_position(
+            position = _find_baseline_insertion_position(
                 reference,
                 working_lines,
             )
